--- a/firebase-firestore/src/androidTest/java/com/google/firebase/firestore/testutil/IntegrationTestUtil.java
+++ b/firebase-firestore/src/androidTest/java/com/google/firebase/firestore/testutil/IntegrationTestUtil.java
@@ -38,7 +38,6 @@ import com.google.firebase.firestore.MetadataChanges;
 import com.google.firebase.firestore.QuerySnapshot;
 import com.google.firebase.firestore.auth.User;
 import com.google.firebase.firestore.core.DatabaseInfo;
-import com.google.firebase.firestore.local.Persistence;
 import com.google.firebase.firestore.model.DatabaseId;
 import com.google.firebase.firestore.testutil.provider.FirestoreProvider;
 import com.google.firebase.firestore.util.AsyncQueue;
@@ -97,7 +96,7 @@ public class IntegrationTestUtil {
   private static final Map<FirebaseFirestore, Boolean> firestoreStatus = new HashMap<>();
 
   /** Default amount of time to wait for a given operation to complete, used by waitFor() helper. */
-  private static final long OPERATION_WAIT_TIMEOUT_MS = 30000;
+  private static final long OPERATION_WAIT_TIMEOUT_MS = 300000;
 
   /**
    * Firestore databases can be subject to a ~30s "cold start" delay if they have not been used
@@ -256,8 +255,8 @@ public class IntegrationTestUtil {
     Logger.setLogLevel(logLevel);
 
     // TODO(Overlay): Remove below once this is ready to ship.
-    Persistence.OVERLAY_SUPPORT_ENABLED = true;
-    Persistence.INDEXING_SUPPORT_ENABLED = true;
+    // Persistence.OVERLAY_SUPPORT_ENABLED = true;
+    // Persistence.INDEXING_SUPPORT_ENABLED = true;
 
     Context context = ApplicationProvider.getApplicationContext();
     DatabaseId databaseId = DatabaseId.forDatabase(projectId, DatabaseId.DEFAULT_DATABASE_ID);