--- a/firebase-firestore/src/main/java/com/google/firebase/firestore/FirebaseFirestore.java
+++ b/firebase-firestore/src/main/java/com/google/firebase/firestore/FirebaseFirestore.java
@@ -529,7 +529,7 @@ public class FirebaseFirestore {
   }
 
   @VisibleForTesting
-  AsyncQueue getAsyncQueue() {
+  public AsyncQueue getAsyncQueue() {
     return asyncQueue;
   }