_All data are from the average of three runs._

As expected, with overlay the query performance stays constant as the number of local mutations grows; while without overlay, the query performance grows linearly as the number of local mutations grows.

**Without Overlay**

Write 10,000 mutations, each document has 1 mutation: 11614ms
Run 100 queries: **3932ms**

Write 40,000 mutations, each document has 4 mutations: 57618ms
Run 100 queries: **7987ms**

Write 70,000 mutations, each document has 7 mutations: 108599ms
Run 100 queries: **11798ms**

Write 100,000 mutations, each document has 10 mutations: 168217ms
Run 100 queries: **15406ms**


**With Overlay**

Write 10,000 mutations, each document has 1 mutation: 15327ms
Run 100 queries: **4493ms**

Write 40,000 mutations, each document has 4 mutations: 67575ms
Run 100 queries: **4278ms**

Write 70,000 mutations, each document has 7 mutations: 123661ms
Run 100 queries: **4767ms**

Write 100,000 mutations, each document has 10 mutations: 185487ms
Run 100 queries: **4699ms**