@wu-hui 2021-11-23
Note: there is also an increase about 30% on mutation write time, because the SDK now writes more data with local mutations. Things might improve when we remove base mutations, we should re-run this test when they are removed.