@wu-hui 2021-11-23
RAW log attache
[performance.txt](https://github.com/firebase/firebase-android-sdk/files/7585293/performance.txt)
d.